Fix a currency loss bug and clean up after review
A withdrawal into a nearly full inventory destroyed rupees. Vanilla's
insertStack returns whether it placed any items, not whether it placed
them all, and it empties what it took from the stack it was handed. The
code trusted the boolean, so a partial insert silently discarded the
remainder after the balance had already been debited. With a 50 of 64
stack of green rupees and no free slot, a bulk withdrawal of 64 gave
back 14 and destroyed the other 50. It now tests the stack.

RupeeChange.solve had no callers since the pouch replaced the coin
trades ten commits ago, along with the constant and the javadoc that
existed only to serve it. Removed. What remains is a denomination table
that no longer does any change making, so it is renamed to Rupees. The
payability analysis its javadoc carried is recorded in the wiki.

Two comments described the pivot as still in progress. One claimed the
loader's volatile guard existed for a method that has not read the
field since that pivot, which would lead an auditor to conclude the
guard was removable.

Also trims comments that restated their own code or repeated the same
explanation across three files.
