+++ src/main/java/net/nieadni/hyliacraft/item/HCDataComponents.java
@@ -9,20 +9,13 @@ import net.minecraft.util.dynamic.Codecs;
 import net.nieadni.hyliacraft.HyliaCraft;
 import net.nieadni.hyliacraft.item.custom.RupeePouchItem;
 
-/**
- * Custom data components for HyliaCraft.
- *
- * <p>Components live on the {@link net.minecraft.item.ItemStack}, not on the player, so a balance travels
- * with the pouch that holds it. That is what lets a player carry several pouches, each with its own money,
- * and have them survive dropping, trading and being stored in a shulker box without any extra work.
- */
+/** Custom data components for HyliaCraft. */
 public class HCDataComponents {
 
     /**
-     * Rupees held in a Rupee Pouch.
-     *
-     * <p>Range-checked at the codec, so a hand-edited or corrupted stack cannot carry a negative or
-     * over-cap balance into the game.
+     * Rupees held in a Rupee Pouch. Lives on the stack, so the balance belongs to the pouch and survives
+     * dropping or being stored. Range-checked at the codec, so a corrupted stack cannot carry a negative
+     * or over-cap balance into the game.
      */
     public static final ComponentType<Integer> RUPEES = Registry.register(
             Registries.DATA_COMPONENT_TYPE,
