+++ src/main/java/net/nieadni/hyliacraft/shop/RupeeCost.java
@@ -9,12 +9,10 @@ import java.util.List;
  *
  * @param item    what the player receives
  * @param cost    the price in rupees
- * @param accepts extra non-rupee items that may be handed over alongside the rupees. Each entry becomes its
- *                own trade offer, because a trade input accepts a single item and not a tag, so "a pumpkin
- *                or a carved pumpkin" cannot be one offer. Empty means the item is bought with rupees alone.
- * @param maxUses purchases allowed from one salesman before the entry shows as out of stock
- * @param restocks whether a spent entry comes back. False makes it a one-off for that salesman's
- *                 lifetime, which is how a pack author gates something rare
+ * @param accepts extra items accepted alongside the rupees. Each becomes its own shop row rather than all
+ *                being required together, so "a pumpkin or a carved pumpkin" is offered as two rows
+ * @param maxUses purchases from one salesman before the entry shows as out of stock
+ * @param restocks whether a spent entry comes back; false makes it a one-off for that salesman
  */
 public record RupeeCost(Item item, int cost, List<Item> accepts, int maxUses, boolean restocks) {
 }
