+++ src/main/java/net/nieadni/hyliacraft/shop/ShopEntry.java
@@ -15,9 +15,7 @@ import java.util.List;
  * <p>Prices come from server-side datapacks, which are never sent to clients, so the list travels with the
  * screen when it opens. Only what the client draws is included; stock rules stay on the server.
  *
- * @param accepted the extra item this row wants alongside the rupees, or {@link Items#AIR} for none. A
- *                 price entry that accepts several items becomes one row per item, so "a pumpkin or a
- *                 carved pumpkin" reads as two offers rather than hiding one of them.
+ * @param accepted the extra item this row wants alongside the rupees, or {@link Items#AIR} for none
  */
 public record ShopEntry(Item item, int cost, Item accepted) {
